fix: 예약내역 페이지 레이아웃 컴포넌트로 분리
